requirements.txt
requests
azure-functions
azure-identity
azure-storage-file-datalake
azure.storage.blob
